requirements.txt
# Azure SDKs
openai>=1.12.0
azure-search-documents>=11.4.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0

# Document processing
pypdf>=4.0.0
python-docx>=1.1.0

# API and UI
fastapi>=0.109.0
uvicorn>=0.27.0
streamlit>=1.30.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.0
rich>=13.7.0
pydantic>=2.5.0
numpy>=1.26.0
